+++ onsiteprint-plugin/blocks/event/block-script-parts/modals/modal-create-participant.js
@@ -1,8 +1,8 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Modal - Create Participant
  *  Creating the Create Participant Content to the Modal in the Event Block.
- ?  Updated: 2025-08-04 - 04:31 (Y:m:d - H:i)
- ?  Info: Changed how the Modal Toggle Listener is used.
+ ?  Updated: 2025-08-06 - 04:15 (Y:m:d - H:i)
+ ?  Info: Added new Modal Clear Form Listener.
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
 ---------------------------------------------------------------------------
@@ -17,7 +17,7 @@
 import * as opModuleBasic from '../../../../assets/js/inc/basic.js'
 import { opModalToggleListener } from '../../../../assets/js/inc/modal/toggle-modal-listener.js'
 import { opColumnInputListener, opCreateParticipantListener, opDownloadCSVFileListener } from '../participant-listeners.js'
-import { opModalClearForm } from '../modals/modal-clear-form.js'
+import { opModalClearFormListener } from './modal-clear-form-listener.js'
 
 
 /* ------------------------------------------------------------------------
@@ -35,22 +35,25 @@ export function opModalCreateParticipant( debug, block, eventId ) {
         if ( debug !== true ) debug = false
         if ( debug ) console.group( `${ functionName }()` )
 
-        ///// Get the Create Participant Template Element.
+        ///// Get the Create Participant Template Elements.
         let modalTemplateElement = block.querySelector( `[id$="-create-participant-template"]` )
-        let modalElement = modalTemplateElement.content.cloneNode(true)
-        let modalHeader = modalElement.querySelector( '.op-header-content__inner' )
-        let modalMain = modalElement.querySelector( '.op-modal-content__inner' )
+        let modal = modalTemplateElement.content.cloneNode(true)
+        let modalHeader = modal.querySelector( '.op-header-content__inner' )
+        let modalMain = modal.querySelector( '.op-modal-content__inner' )
         let formElement = modalMain.querySelector( '.op-form' )
         let columnInputElements = formElement.querySelectorAll( '.op-col-input input' )
+        let saveButton = modal.querySelector( '.op-button-save' )
+        let cancelButton = modal.querySelector( '.op-cancel_create-participant' )
+
+        ///// Get the Modal Elements.
         let addButton = block.querySelector( '.op-button-add' )
-        let cancelButton = block.querySelector( '.op-button-cancel' )
-        let saveButton = modalMain.querySelector( '.op-button-save' )
+        let modalElement = block.querySelector( '.op-modal' )
 
         ///// Set Modal Toggle Listener to the Add Participant Button.
         opModalToggleListener( debug, addButton, true, modalHeader, modalMain ) 
 
-        ///// Set Modal Toggle Listener to the Modal Close Button.
-        opModalToggleListener( debug, cancelButton, false )
+        ///// Set Modal Clear Form Listener to the Modal Close Button.
+        opModalClearFormListener( debug, modalElement, cancelButton )
 
         ///// Set Column Input Listener to All of the Column Input Elements in the Modal Form.
         columnInputElements.forEach( inputElement => {